refactor processMessages
